// generated/openweather-schemas/currentWeatherResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * OpenWeatherMap API
 * Weather data API
 * OpenAPI spec version: 2.5.0
 */
import type { Coord } from "./coord";
import type { Weather } from "./weather";
import type { Main } from "./main";
import type { Wind } from "./wind";
import type { Clouds } from "./clouds";
import type { Sys } from "./sys";

export interface CurrentWeatherResponse {
  coord?: Coord;
  weather?: Weather[];
  base?: string;
  main?: Main;
  visibility?: number;
  wind?: Wind;
  clouds?: Clouds;
  /** Time of data calculation, unix, UTC */
  dt?: number;
  sys?: Sys;
  timezone?: number;
  id?: number;
  name?: string;
  cod?: number;
}